Move FixedLengthSource to ktx module
