+++ instances/android/app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
             android:label="@string/app_name"
             android:name="com.flipperdevices.app.FlipperApplication"
             android:supportsRtl="false"
+            android:localeConfig=""
             android:theme="@style/FlipperTheme">
         <activity
                 android:exported="true"
